Fix #593, change OCS_off_t to a signed type
The C library type is signed, and this makes the
result check work as intended.

+++ src/unit-test-coverage/ut-stubs/inc/OCS_sys_types.h
@@ -32,7 +32,7 @@
 /* types normally defined in sys/types.h */
 /* ----------------------------------------- */
 typedef ptrdiff_t       OCS_ssize_t;
-typedef size_t          OCS_off_t;
+typedef long            OCS_off_t;
 typedef unsigned int    OCS_mode_t;
 typedef long            OCS_time_t;
 typedef int             OCS_pid_t;
